requirements.txt
#must use python 3.11
robotpy-ntcore